feat: 나의 여행 화면 완성

+++ feature/travel-init/src/main/java/com/jun/tripguide_v2/feature/travelInit/mapper/StartingPointMapper.kt
@@ -5,5 +5,5 @@ import com.jun.tripguide_v2.core.model.StartingPoint
 fun String?.toStartingPoint(): StartingPoint? {
     if (isNullOrBlank()) return null
     val (name, x, y) = split("/")
-    return StartingPoint(name, x, y)
+    return StartingPoint(name, x.toDouble(), y.toDouble())
 }
